備份/requirements.txt
# WebTech Automation System Dependencies

# Core dependencies
selenium>=4.15.0
webdriver-manager>=4.0.0
pillow>=10.0.0
openai>=1.0.0

# Optional dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Line Bot related (if needed)
# line-bot-sdk>=3.5.0
# flask>=2.3.0
# gunicorn>=21.0.0

# Logging and monitoring (optional)
# loguru>=0.7.0
# sentry-sdk>=1.32.0

# Image processing enhancement (optional)
# opencv-python>=4.8.0
# numpy>=1.24.0